Inflate a scalar to an (m x n) matrix with identical entries

Used in elementwise group operations Ops.ratm to *inflate* a scalar ratm object such that it can be, e.g., elementwise added. It is used after upgrade_objects is called and therefore only relevant for

polm,

lpolm,

stsp,

pseries,

zvalues

In other words, one cannot inflate lmfd or rmfd objects.

Usage

inflate_object(e, m, n)

Arguments

e

Object of type polm, lpolm, stsp, pseries, or zvalues

m

Integer. Output dimension to which the scalar should be inflated.

n

Integer. Output dimension to which the scalar should be inflated.

Value

Object of same type but with *inflated* dimension

Examples

(lp = test_lpolm(degree_max = 1, degree_min = -1))
#> ( 1 x 1 ) Laurent polynomial matrix with degree <= 1, and minimal degree >= -1
#>       z^-1 [,1] z^0 [,1] z^1 [,1]
#> [1,] -0.1508111  1.53411 1.938096
inflate_object(lp, 2, 2)
#> ( 2 x 2 ) Laurent polynomial matrix with degree <= 1, and minimal degree >= -1
#>       z^-1 [,1]       [,2] z^0 [,1]    [,2] z^1 [,1]     [,2]
#> [1,] -0.1508111 -0.1508111  1.53411 1.53411 1.938096 1.938096
#> [2,] -0.1508111 -0.1508111  1.53411 1.53411 1.938096 1.938096

(ss = test_stsp(s = 1))
#> statespace realization [1,1] with s = 1 states
#>            s[1]       u[1]
#> s[1]  1.0721054 -0.9162056
#> x[1] -0.1784693  1.0000000
inflate_object(ss, 2, 2)
#> statespace realization [2,2] with s = 2 states
#>            s[1]       s[2]       u[1]       u[2]
#> s[1]  1.0721054  0.0000000 -0.9162056 -0.9162056
#> s[2]  0.0000000  1.0721054 -0.9162056 -0.9162056
#> x[1] -0.1784693  0.0000000  1.0000000  1.0000000
#> x[2]  0.0000000 -0.1784693  1.0000000  1.0000000