Constructor for Statespace Realizations

Any rational (\((m,n)\)-dimensional) matrix \(x(z)\) which has no pole at \(z=0\) may be represented as $$x(z) = C(z^{-1} I_s - A)^{-1} B + D$$ Here \(I_s\) denotes the \((s,s)\)-dimensional identity matrix and \(A,B,C,D\) are (real or complex valued) matrices of size \((s,s)\), \((s,n)\), \((m,s)\) and \((m,n)\) respectively. The integer \(s \geq 0\) is called the state dimension of the above statespace realization of \(x(z)\).

Usage

stsp(A, B, C, D)

Arguments

A

\((s,s)\) matrix (or a vector of length \((s^2)\))

B

\((s,n)\) matrix (or a vector of length \((sn)\))

C

\((m,s)\) matrix (or a vector of length \((ms)\))

D

\((m,n)\) matrix (or a vector of length \((mn)\))

Value

An object of class stsp.

Details

Internally statespace realizations are stored as an \((s+m, s+n)\)-dimensional matrix with an attribute order = c(m,n,s) and a class attribute c("stsp","ratm").

Any of the integers \(m,n,s\) may be zero.

If the arguments A,B,C are missing, then a statespace realization with statespace dimension \(s=0\) is constructed. In this case D must a be matrix or a scalar. If A,B,C are given (and compatible) and D is missing, then D = diag(x = 1, nrow = m, ncol = n) is used.

See also

Examples

### x(z) =  I
stsp(D = diag(3))
#> statespace realization [3,3] with s = 0 states
#>      u[1] u[2] u[3]
#> x[1]    1    0    0
#> x[2]    0    1    0
#> x[3]    0    0    1

### random (2 x 3) rational matrix in statespace form with state dimension s = 4
x = stsp(A = stats::rnorm(4*4), B = stats::rnorm(4*3), C = stats::rnorm(2*4))

is.stsp(x)
#> [1] TRUE
dim(x)
#> m n s 
#> 2 3 4 
str(x)
#> ( 2 x 3 ) statespace realization with s = 4 states
print(x, digits = 3)
#> statespace realization [2,3] with s = 4 states
#>        s[1]   s[2]   s[3]   s[4]   u[1]   u[2]   u[3]
#> s[1] -1.582  0.847 -0.084 -0.812  0.564 -0.001  0.667
#> s[2] -1.748  0.147  0.413 -0.577 -0.891  0.070 -0.846
#> s[3]  0.564  1.440  0.160 -1.805 -0.444 -1.162  0.096
#> s[4]  0.140  0.279  1.310  0.024 -0.137  1.001 -2.402
#> x[1]  0.660 -0.547 -0.667  1.223  1.000  0.000  0.000
#> x[2] -0.496 -0.230 -0.974 -1.191  0.000  1.000  0.000