Rational Matrix Transpose

Compute the transpose of a rational matrix \(x(z)\).

Usage

# S3 method for polm
t(x)

# S3 method for lpolm
t(x)

# S3 method for lmfd
t(x)

# S3 method for rmfd
t(x)

# S3 method for stsp
t(x)

# S3 method for pseries
t(x)

# S3 method for zvalues
t(x)

Arguments

x

rational matrix object, i.e. a polm, lpolm, lmfd, rmfd, stsp, pseries or zvalues object.

Value

A rational matrix object, which represents the transposed rational matrix \(x'(z)\). The output is of the same class as the input x unless

x is an lmfd or an rmfd object: The transposition of an lmfd object is an rmfd object, and vice versa.

Examples

x = test_polm(dim = c(2,3), degree = 3, random = TRUE)
all.equal(pseries(t(x)), t(pseries(x)))
#> [1] TRUE

x = test_stsp(dim = c(3,2), s = 1)
all.equal(zvalues(t(x)), t(zvalues(x)))
#> [1] TRUE

# the transpose of an LMFD object is RMFD
(x = test_lmfd(dim = c(3,2), degrees = c(1,1)))
#> ( 3 x 2 ) left matrix fraction description a^(-1)(z) b(z) with degrees (p = 1, q = 1)
#> left factor a(z):
#>      z^0 [,1]  [,2]  [,3]  z^1 [,1]       [,2]       [,3]
#> [1,]        1     0     0 1.3249976  1.7078150 -1.1669165
#> [2,]        0     1     0 1.5178235 -2.3328057 -0.1161027
#> [3,]        0     0     1 0.7004871 -0.9891472 -0.8800032
#> right factor b(z):
#>       z^0 [,1]       [,2]   z^1 [,1]        [,2]
#> [1,] -1.081509 -0.7079270 -0.5170237  0.18757343
#> [2,]  1.096239  0.8248984  0.2090721 -0.14121976
#> [3,]  1.755518  0.8607571 -0.1762874 -0.00452162
t(x)
#> ( 2 x 3 ) right matrix fraction description d(z) c^(-1)(z) with degrees deg(c(z)) = p = 1, deg(d(z)) = q = 1
#> left factor d(z):
#>       z^0 [,1]      [,2]      [,3]   z^1 [,1]       [,2]        [,3]
#> [1,] -1.081509 1.0962393 1.7555177 -0.5170237  0.2090721 -0.17628739
#> [2,] -0.707927 0.8248984 0.8607571  0.1875734 -0.1412198 -0.00452162
#> right factor c(z):
#>      z^0 [,1]  [,2]  [,3]  z^1 [,1]       [,2]       [,3]
#> [1,]        1     0     0  1.324998  1.5178235  0.7004871
#> [2,]        0     1     0  1.707815 -2.3328057 -0.9891472
#> [3,]        0     0     1 -1.166916 -0.1161027 -0.8800032
all.equal(x, t(t(x)))
#> [1] TRUE