Constructor for Polynomial Matrices

polm objects represent polynomial matrices $$a(z) = a_0 + a_1 z + \cdots + a_p z^p$$ If the matrix \(a(z)\) is an \((m,n)\) dimensional polynomial matrix (i.e. the coefficients \(a_i\) are \((m,n)\) dimensional real or complex valued matrices) then the polm object stores the coefficients in an (m,n,p+1) dimensional (numeric or complex) array together with a class attribute c("polm","ratm").
The constructor function polm(a) takes a (numeric or complex) vector, matrix or 3-dimensional array and returns a polm object.

Usage

polm(a)

Arguments

a

either a (numeric or complex) vector, matrix or 3-D array. A vector is coerced to a scalar (i.e. \((1,1)\)-dimensional) polynomial and a matrix gives a polynomial matrix of zero degree.

Value

An object of class polm.

Details

Any of the dimensions of the 3-dimensional array may also be zero. In particular, if the third dimension is zero, then the polm object is interpreted as the zero polynomial.

For important methods and functions for this class have a look at the "see also" section.

See also

Examples

# (1 x 1) polynomial matrix a(z) =  0 + 1z + 2z^2
polm(0:2)
#> ( 1 x 1 ) matrix polynomial with degree <= 2 
#>      z^0 [,1] z^1 [,1] z^2 [,1]
#> [1,]        0        1        2

# (2 x 3) polynomial matrix a(z) = a0 (degree is zero)
polm(diag(1, nrow = 2, ncol = 3))
#> ( 2 x 3 ) matrix polynomial with degree <= 0 
#>      z^0 [,1]  [,2]  [,3]
#> [1,]        1     0     0
#> [2,]        0     1     0

# random (2 x 3) polynomial matrix a(z) = a0 + a1 z + a2 z^2 + a3 z^3 (degree = 3)
polm(array(stats::rnorm(2*3*4), dim = c(2,3,4)))
#> ( 2 x 3 ) matrix polynomial with degree <= 3 
#>       z^0 [,1]     [,2]       [,3]  z^1 [,1]       [,2]       [,3]   z^2 [,1]
#> [1,] 0.6370071 2.137362 -0.3834814 0.6609966 -0.1441613 -0.4183031 0.62482433
#> [2,] 0.5058216 0.869301  1.0289366 1.4180023  1.1997984  1.2666229 0.08006296
#>            [,2]      [,3]    z^3 [,1]      [,2]       [,3]
#> [1,] -0.3267119 -1.022723  0.54688930 0.2083001 -0.7496543
#> [2,]  0.2870435  0.459892 -0.02820572 0.6123943  0.8461007

# random (2 x 1) polynomial matrix with complex coefficients (degree = 2)
a = polm(array(complex(real = stats::rnorm(2*1*3), 
               imaginary = stats::rnorm(2*1*3)), dim = c(2,1,3)))
is.polm(a)
#> [1] TRUE
dim(a)
#> m n p 
#> 2 1 2 
str(a)
#> ( 2 x 1 ) matrix polynomial with degree <= 2
print(a, digits = 3)
#> ( 2 x 1 ) matrix polynomial with degree <= 2 
#>          z^0 [,1]      z^1 [,1]      z^2 [,1]
#> [1,] 0.532+0.578i -1.514-1.296i  0.372-0.958i
#> [2,] 0.516+0.903i -0.288-2.804i -0.820-0.166i