Create Test Laurent Polynomial Matrix

See test_polm. This simple tool creates Laurent polynomial matrices for testing purposes. The column_start_matrix, i.e. the matrix consisting of the columns pertaining to the smallest column degrees, the value_at_0, and that column_end_matrix matrix are set in this order. Thus, column_end_matrix may potentially overwrite value_at_0.

Usage

test_lpolm(
  dim = c(1, 1),
  degree_max = 1,
  degree_min = -1,
  random = FALSE,
  col_start_matrix = NULL,
  value_at_0 = NULL,
  col_end_matrix = NULL
)

Arguments

dim

two dimensional vector of non negative integers, determines the dimension of the polynomial matrix to be created. If the prescribed number of rows or number of columns is zero then an "empty" polynomial matrix is generated. In this case all parameters below are ignored.

degree_max, degree_min

Integer, vector, or matrix specifying minimal and maximal (column) degrees

random

If TRUE the coefficents are generated by drawing from a normal distribution. If FALSE then the coefficient of the \(k\)-th power \(z^k\) of the (i,j)-th entry is set equal to "\(ijk\)". In this case the parameters below are ignored!

col_start_matrix

Force a matrix for lowest column degrees

value_at_0

desired value of the polynomial at \(z=0\) (or NULL).

col_end_matrix

desired column end matrix (or NULL).

Value

lpolm object

Details

In contrast to test_polm, the absolute value of the zeros cannot be forced to be larger than a certain lower bound bzeroes.

Examples

(lp = test_lpolm(dim = c(1,1), degree_max = 1, degree_min = -2))
#> ( 1 x 1 ) Laurent polynomial matrix with degree <= 1, and minimal degree >= -2
#>      z^-2 [,1]  z^-1 [,1]  z^0 [,1]  z^1 [,1]
#> [1,]  1.371118 -0.6406476 0.2683996 0.9329544
(lp = test_lpolm(dim = c(3,3), degree_max = c(0,1,2), degree_min = -2))
#> ( 3 x 3 ) Laurent polynomial matrix with degree <= 2, and minimal degree >= -2
#>       z^-2 [,1]       [,2]      [,3] z^-1 [,1]       [,2]       [,3]  z^0 [,1]
#> [1,]  0.3845852  0.7583993 0.8589292  0.552901 -1.1838369  0.2791122 0.3298084
#> [2,] -1.3915363 -1.3118794 0.9971988 -1.256305  0.5915774 -1.7356670 1.3404775
#> [3,] -0.6812212  0.3931899 0.3328105  1.594484  0.6505731 -0.1929422 0.2707204
#>            [,2]        [,3] z^1 [,1]      [,2]       [,3] z^2 [,1]  [,2]
#> [1,]  0.7078426 -0.62144041        0  1.502063 -0.3018244        0     0
#> [2,] -0.6580599 -1.19847885        0 -0.524480 -0.2292870        0     0
#> [3,] -1.7006358 -0.02689433        0  1.223345 -0.2472679        0     0
#>            [,3]
#> [1,]  1.5363339
#> [2,]  0.5624288
#> [3,] -1.1651476
(lp = test_lpolm(dim = c(3,3), degree_max = 1, degree_min = c(0,-1,-2)))
#> ( 3 x 3 ) Laurent polynomial matrix with degree <= 1, and minimal degree >= -2
#>      z^-2 [,1]  [,2]      [,3] z^-1 [,1]       [,2]       [,3]   z^0 [,1]
#> [1,]         0     0 0.5379363         0 -0.4893129 -0.3499187  0.5532784
#> [2,]         0     0 1.2741732         0  0.1864930 -1.1311173 -0.8006727
#> [3,]         0     0 1.2260134         0  0.9998520 -1.0004994  1.1954122
#>            [,2]       [,3]     z^1 [,1]       [,2]       [,3]
#> [1,] -0.9115983  0.8310492 -0.811518898  0.2564780  0.4536129
#> [2,] -0.8523967 -0.1263899  0.001488695 -0.2580758  1.3780257
#> [3,]  0.4711564  2.8058865  1.224060479  0.5334909 -0.1151941